Partial functionality update
Bold button should function. When the user clicks on something on the left column, it will toggle "custom" immediately. So far this is enabled for all the features on the left. Additionally, when the "bold" toggle is turned off, it will maintain the state of the "spacing" toggle. What is still needed is the opposite case where the spacing is turned off, both with and without the bold setting being turned on. The specific blockage is actually figuring out how to revert the spacing changes, without a full reload. Also need to add mutual exclusivity still.