Mutual exclusivity
Mostly implemented.
Blockage: I needed to remove the effects of custom and preset for the purpose of mutual exclusivity and found difficulty due to our use of reload();.
Solution: Added a function that will remove a preset and implemented it.

Known bugs: Interacting with anything on the left with a preset enabled will disable the preset. This half works, but the bugs are that some of the values are remembered when the left interaction is implemented. This leads to, in some cases, text being successively bolded until everything is bold. I believe this error is due to flagging issues for some of the user cases and I'm going to keep looking into it.